Split invoke into separate header
